+++ sushii-2/migrations/20210204013144_alter-guild-config-disabled-channels.sql
@@ -1,0 +1,2 @@
+ALTER TABLE guild_configs
+ADD COLUMN disabled_channels BIGINT [];