+++ sushii-2/src/commands/settings/mod.rs
@@ -1,13 +1,14 @@
 use serenity::framework::standard::macros::group;
 
 mod default;
+mod disable_channel;
 mod list;
 mod mute;
 
-use self::{default::*, list::*, mute::*};
+use self::{default::*, disable_channel::*, list::*, mute::*};
 
 #[group]
-#[commands(list, mute)]
+#[commands(list, mute, disablechannel, enablechannel, disabledchannels)]
 #[description("Guild settings, requires MANAGE_GUILD permissions")]
 #[prefix("settings")]
 #[only_in("guild")]