+++ sushii-2/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -  `deletecase` command to delete a mod log case
+-  `settings disablechannel` `settings enablechannel` `settings disabledchannels` commands
 
 ## [0.2.0-rc.3] - 2021-02-02